Fixed head circumference validation. Restored compatibility with features "add patient", "add parent".
