+++ MedicalDocManagment.WebUI/app/components/childrens-card/child-card/speech-therapists-examination/form.component.css
@@ -1,0 +1,24 @@
+.hyphens {
+  -webkit-hyphens: auto;
+  -moz-hyphens: auto;
+  -ms-hyphens: auto;
+  hyphens: auto;
+}
+.no-horizontal-padding {
+  padding-left: 0px;
+  padding-right: 0px;
+}
+textarea {
+  resize: vertical;
+}
+#neurologicalStateHeadDataHeadText {
+  width: 25%;
+}
+#neurologicalStateHeadCircumferenceAlert {
+  margin-left: 25%;
+}
+:host /deep/ .tooltip-inner {
+  margin-left: 25%;
+  max-width: 100%;
+  text-align: left;
+}
