+++ MedicalDocManagment.WebUI/app/components/childrens-card/parent/child-card-add-parent.component.ts
@@ -4,31 +4,31 @@ import { Observable } from 'rxjs/Observable';
 
 declare var jQuery: any;
 
-import ChildCardService from "../../../services/child-card.service";
+import ChildrenCardService from "../../../services/children-card.service";
 import ParentModel from "../../../models/child-card/parent.model";
 
 @Component({
     moduleId: module.id,
     selector: 'child-card-add-parent',
     templateUrl: 'child-card-add-parent.component.html',
-    providers: [ChildCardService],
+    providers: [ChildrenCardService],
     styleUrls: ['child-card-add-parent.component.css']
 })
 
 export default class ChildCardAddParentComponent {
     @Output() parentAddedEvent: EventEmitter<ParentModel>;
 
-    private _childCardService: ChildCardService;
+    private _childrenCardService: ChildrenCardService;
     private _isAdding: boolean;
     private _isErrorOnAdding: boolean;
     private _lastErrorMessage: string;
     private _parent: ParentModel;
     private _phoneMask: Array<string | RegExp>;
 
-    constructor(childCardService: ChildCardService) {
+    constructor(childrenCardService: ChildrenCardService) {
         this.parentAddedEvent = new EventEmitter<ParentModel>();
 
-        this._childCardService = childCardService;
+        this._childrenCardService = childrenCardService;
         this._isAdding = false;
         this._isErrorOnAdding = false;
         this._lastErrorMessage = null;
@@ -39,7 +39,7 @@ export default class ChildCardAddParentComponent {
     private _onSave(form: FormGroup): void {
         this._isAdding = true;
         this._isErrorOnAdding = false;
-        this._childCardService.addParent(this._parent)
+        this._childrenCardService.addParent(this._parent)
             .subscribe((result: any) => {
                 if (result) {
                     this._isAdding = false;
